+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 - Update scythe textures
+- Netherite scythes no longer burn
+- Obese crops can now be sliced
 - Added flowering oaks, which will grow apples, and they also have an interesting log
+- Change foliage block voxel shape